:recycle: refactor update and complete functionality from AssessmentService into AssessmentUpdateService make various small changes suggested by IDE warnings
