+++ src/main/kotlin/uk/gov/justice/digital/assessments/restclient/assessmentupdateapi/CreateAssessmentResponse.kt
@@ -1,3 +1,3 @@
-package uk.gov.justice.digital.hmpps.offenderassessmentsupdates.api
+package uk.gov.justice.digital.assessments.restclient.assessmentupdateapi
 
 data class CreateAssessmentResponse(val oasysSetPk: Long? = null)
